Connections to Database/SQL + HTML
In Progress.

Currently working on making html files so that the program can output updated data into a directory for the webhost.

Currently working on connecting the SQL Database to the program so the proper data can be retreved and used to populate the HTML files.

Will also need to accept data from the webhost and send it to the proper tables within the database.