Database and updated connector
The database that includes the testing data has been added here. You can open and edit this on your pc by installing MySQL Workbench and placing this in C:\ProgramData\MySQL\MySQL server (your version number here)\data\mydatabase\(daystruct_database folder goes here)

Then you should be able to open it when you start your localhost server and will be able to see and edit the queries.

The database connector in the c++ driver connects to this database but is not complete. I am still working through using the MySQL functions that came in the (already included in the project) MySQL header files to send and collect data through the database.